refactor: separate place holder and polling runner
